+++ Client/EntryPoint/EntryPoint.vcxproj
@@ -153,11 +153,11 @@
     <ClCompile Include="main.cpp" />
   </ItemGroup>
   <ItemGroup>
-    <ProjectReference Include="..\..\Engines\GraphicsEngine\GraphicsEngine.vcxproj">
-      <Project>{2d7a80f7-cb5d-4515-8789-5f84e6c14627}</Project>
+    <ProjectReference Include="..\..\Engines\Graphics\Graphics.vcxproj">
+      <Project>{386feae8-9910-4816-b846-06633983f3b2}</Project>
     </ProjectReference>
-    <ProjectReference Include="..\..\Engines\PhysicsEngine\PhysicsEngine.vcxproj">
-      <Project>{cb82332b-74c2-4c23-9760-385cade2ce6a}</Project>
+    <ProjectReference Include="..\..\Engines\Physics\Physics.vcxproj">
+      <Project>{1a6b1b96-89c0-46a7-a6bf-3d43298d7e26}</Project>
     </ProjectReference>
     <ProjectReference Include="..\..\Engines\RenderTools\RenderTools.vcxproj">
       <Project>{f403ed3b-ea47-415b-b9c9-4ba159b063d8}</Project>
