+++ Engines/Graphics/Graphics.vcxproj
@@ -20,8 +20,8 @@
   </ItemGroup>
   <PropertyGroup Label="Globals">
     <VCProjectVersion>15.0</VCProjectVersion>
-    <ProjectGuid>{2D7A80F7-CB5D-4515-8789-5F84E6C14627}</ProjectGuid>
-    <RootNamespace>GraphicsEngine</RootNamespace>
+    <ProjectGuid>{386FEAE8-9910-4816-B846-06633983F3B2}</ProjectGuid>
+    <RootNamespace>Graphics</RootNamespace>
     <WindowsTargetPlatformVersion>10.0.15063.0</WindowsTargetPlatformVersion>
   </PropertyGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.Default.props" />
@@ -81,6 +81,9 @@
       <AdditionalDependencies>MGL.lib;</AdditionalDependencies>
       <AdditionalLibraryDirectories>..\..\..\MGL\x64\Debug;</AdditionalLibraryDirectories>
     </Lib>
+    <ProjectReference>
+      <LinkLibraryDependencies>true</LinkLibraryDependencies>
+    </ProjectReference>
   </ItemDefinitionGroup>
   <ItemDefinitionGroup Condition="'$(Configuration)|$(Platform)'=='Debug|Win32'">
     <ClCompile>
@@ -93,6 +96,9 @@
       <AdditionalDependencies>MGL.lib;</AdditionalDependencies>
       <AdditionalLibraryDirectories>..\..\..\MGL\Debug;</AdditionalLibraryDirectories>
     </Lib>
+    <ProjectReference>
+      <LinkLibraryDependencies>true</LinkLibraryDependencies>
+    </ProjectReference>
   </ItemDefinitionGroup>
   <ItemDefinitionGroup Condition="'$(Configuration)|$(Platform)'=='Release|Win32'">
     <ClCompile>
@@ -111,6 +117,9 @@
       <AdditionalDependencies>MGL.lib;</AdditionalDependencies>
       <AdditionalLibraryDirectories>..\..\..\MGL\Release;</AdditionalLibraryDirectories>
     </Lib>
+    <ProjectReference>
+      <LinkLibraryDependencies>true</LinkLibraryDependencies>
+    </ProjectReference>
   </ItemDefinitionGroup>
   <ItemDefinitionGroup Condition="'$(Configuration)|$(Platform)'=='Release|x64'">
     <ClCompile>
@@ -129,18 +138,10 @@
       <AdditionalDependencies>MGL.lib;</AdditionalDependencies>
       <AdditionalLibraryDirectories>..\..\..\MGL\x64\Release;</AdditionalLibraryDirectories>
     </Lib>
+    <ProjectReference>
+      <LinkLibraryDependencies>true</LinkLibraryDependencies>
+    </ProjectReference>
   </ItemDefinitionGroup>
-  <ItemGroup>
-    <ClInclude Include="MGL_AdditionalStructures.hpp" />
-    <ClInclude Include="MGL_Window.hpp" />
-    <ClInclude Include="MyGraphicsLibraryEngine.hpp" />
-  </ItemGroup>
-  <ItemGroup>
-    <ClCompile Include="MGL_HUDRender.cpp" />
-    <ClCompile Include="MGL_MapRender.cpp" />
-    <ClCompile Include="MGL_QueueRender.cpp" />
-    <ClCompile Include="MyGraphicsLibraryEngine.cpp" />
-  </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\..\Objects\AbstractObjects\AbstractObjects.vcxproj">
       <Project>{17253703-7209-43b3-863b-0a8fe5d524eb}</Project>
@@ -154,6 +155,17 @@
     <ProjectReference Include="..\RenderTools\RenderTools.vcxproj">
       <Project>{f403ed3b-ea47-415b-b9c9-4ba159b063d8}</Project>
     </ProjectReference>
+  </ItemGroup>
+  <ItemGroup>
+    <ClInclude Include="MGL_AdditionalStructures.hpp" />
+    <ClInclude Include="MGL_Window.hpp" />
+    <ClInclude Include="MyGraphicsLibraryEngine.hpp" />
+  </ItemGroup>
+  <ItemGroup>
+    <ClCompile Include="MGL_HUDRender.cpp" />
+    <ClCompile Include="MGL_MapRender.cpp" />
+    <ClCompile Include="MGL_QueueRender.cpp" />
+    <ClCompile Include="MyGraphicsLibraryEngine.cpp" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
   <ImportGroup Label="ExtensionTargets">
