+++ SkyMages.sln
@@ -19,10 +19,6 @@ Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "EntryPoint", "Client\EntryP
 EndProject
 Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "GameWindow", "Client\GameWindow\GameWindow.vcxproj", "{51ACE45F-329D-405E-9AE8-913A942C997C}"
 EndProject
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "GraphicsEngine", "Engines\GraphicsEngine\GraphicsEngine.vcxproj", "{2D7A80F7-CB5D-4515-8789-5F84E6C14627}"
-EndProject
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "PhysicsEngine", "Engines\PhysicsEngine\PhysicsEngine.vcxproj", "{CB82332B-74C2-4C23-9760-385CADE2CE6A}"
-EndProject
 Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "RenderTools", "Engines\RenderTools\RenderTools.vcxproj", "{F403ED3B-EA47-415B-B9C9-4BA159B063D8}"
 EndProject
 Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "LogicEngine", "Objects\LogicEngine\LogicEngine.vcxproj", "{ACE8714C-1DDE-4D4D-B947-092ACB940888}"
@@ -38,6 +34,10 @@ EndProject
 Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "EquipableItems", "Objects\EquipableItems\EquipableItems.vcxproj", "{DFDDC18D-7CDB-4842-94B5-5F7332AC95EC}"
 EndProject
 Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "ObjectQueue", "Objects\ObjectQueue\ObjectQueue.vcxproj", "{94B05EC3-2FE9-42D9-B765-A322DF2C87CD}"
+EndProject
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Physics", "Engines\Physics\Physics.vcxproj", "{1A6B1B96-89C0-46A7-A6BF-3D43298D7E26}"
+EndProject
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Graphics", "Engines\Graphics\Graphics.vcxproj", "{386FEAE8-9910-4816-B846-06633983F3B2}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -79,22 +79,6 @@ Global
 		{51ACE45F-329D-405E-9AE8-913A942C997C}.Release|x64.Build.0 = Release|x64
 		{51ACE45F-329D-405E-9AE8-913A942C997C}.Release|x86.ActiveCfg = Release|Win32
 		{51ACE45F-329D-405E-9AE8-913A942C997C}.Release|x86.Build.0 = Release|Win32
-		{2D7A80F7-CB5D-4515-8789-5F84E6C14627}.Debug|x64.ActiveCfg = Debug|x64
-		{2D7A80F7-CB5D-4515-8789-5F84E6C14627}.Debug|x64.Build.0 = Debug|x64
-		{2D7A80F7-CB5D-4515-8789-5F84E6C14627}.Debug|x86.ActiveCfg = Debug|Win32
-		{2D7A80F7-CB5D-4515-8789-5F84E6C14627}.Debug|x86.Build.0 = Debug|Win32
-		{2D7A80F7-CB5D-4515-8789-5F84E6C14627}.Release|x64.ActiveCfg = Release|x64
-		{2D7A80F7-CB5D-4515-8789-5F84E6C14627}.Release|x64.Build.0 = Release|x64
-		{2D7A80F7-CB5D-4515-8789-5F84E6C14627}.Release|x86.ActiveCfg = Release|Win32
-		{2D7A80F7-CB5D-4515-8789-5F84E6C14627}.Release|x86.Build.0 = Release|Win32
-		{CB82332B-74C2-4C23-9760-385CADE2CE6A}.Debug|x64.ActiveCfg = Debug|x64
-		{CB82332B-74C2-4C23-9760-385CADE2CE6A}.Debug|x64.Build.0 = Debug|x64
-		{CB82332B-74C2-4C23-9760-385CADE2CE6A}.Debug|x86.ActiveCfg = Debug|Win32
-		{CB82332B-74C2-4C23-9760-385CADE2CE6A}.Debug|x86.Build.0 = Debug|Win32
-		{CB82332B-74C2-4C23-9760-385CADE2CE6A}.Release|x64.ActiveCfg = Release|x64
-		{CB82332B-74C2-4C23-9760-385CADE2CE6A}.Release|x64.Build.0 = Release|x64
-		{CB82332B-74C2-4C23-9760-385CADE2CE6A}.Release|x86.ActiveCfg = Release|Win32
-		{CB82332B-74C2-4C23-9760-385CADE2CE6A}.Release|x86.Build.0 = Release|Win32
 		{F403ED3B-EA47-415B-B9C9-4BA159B063D8}.Debug|x64.ActiveCfg = Debug|x64
 		{F403ED3B-EA47-415B-B9C9-4BA159B063D8}.Debug|x64.Build.0 = Debug|x64
 		{F403ED3B-EA47-415B-B9C9-4BA159B063D8}.Debug|x86.ActiveCfg = Debug|Win32
@@ -159,6 +143,22 @@ Global
 		{94B05EC3-2FE9-42D9-B765-A322DF2C87CD}.Release|x64.Build.0 = Release|x64
 		{94B05EC3-2FE9-42D9-B765-A322DF2C87CD}.Release|x86.ActiveCfg = Release|Win32
 		{94B05EC3-2FE9-42D9-B765-A322DF2C87CD}.Release|x86.Build.0 = Release|Win32
+		{1A6B1B96-89C0-46A7-A6BF-3D43298D7E26}.Debug|x64.ActiveCfg = Debug|x64
+		{1A6B1B96-89C0-46A7-A6BF-3D43298D7E26}.Debug|x64.Build.0 = Debug|x64
+		{1A6B1B96-89C0-46A7-A6BF-3D43298D7E26}.Debug|x86.ActiveCfg = Debug|Win32
+		{1A6B1B96-89C0-46A7-A6BF-3D43298D7E26}.Debug|x86.Build.0 = Debug|Win32
+		{1A6B1B96-89C0-46A7-A6BF-3D43298D7E26}.Release|x64.ActiveCfg = Release|x64
+		{1A6B1B96-89C0-46A7-A6BF-3D43298D7E26}.Release|x64.Build.0 = Release|x64
+		{1A6B1B96-89C0-46A7-A6BF-3D43298D7E26}.Release|x86.ActiveCfg = Release|Win32
+		{1A6B1B96-89C0-46A7-A6BF-3D43298D7E26}.Release|x86.Build.0 = Release|Win32
+		{386FEAE8-9910-4816-B846-06633983F3B2}.Debug|x64.ActiveCfg = Debug|x64
+		{386FEAE8-9910-4816-B846-06633983F3B2}.Debug|x64.Build.0 = Debug|x64
+		{386FEAE8-9910-4816-B846-06633983F3B2}.Debug|x86.ActiveCfg = Debug|Win32
+		{386FEAE8-9910-4816-B846-06633983F3B2}.Debug|x86.Build.0 = Debug|Win32
+		{386FEAE8-9910-4816-B846-06633983F3B2}.Release|x64.ActiveCfg = Release|x64
+		{386FEAE8-9910-4816-B846-06633983F3B2}.Release|x64.Build.0 = Release|x64
+		{386FEAE8-9910-4816-B846-06633983F3B2}.Release|x86.ActiveCfg = Release|Win32
+		{386FEAE8-9910-4816-B846-06633983F3B2}.Release|x86.Build.0 = Release|Win32
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -168,8 +168,6 @@ Global
 		{C5C17A7E-3702-464B-A46A-39679106B95B} = {96371A14-F893-4DC5-96BA-DEA333709AFD}
 		{C1110F89-0B60-42C5-B6B2-B7571D6FB743} = {96371A14-F893-4DC5-96BA-DEA333709AFD}
 		{51ACE45F-329D-405E-9AE8-913A942C997C} = {96371A14-F893-4DC5-96BA-DEA333709AFD}
-		{2D7A80F7-CB5D-4515-8789-5F84E6C14627} = {785A6394-7DDC-42FB-A7ED-2898EDC177EE}
-		{CB82332B-74C2-4C23-9760-385CADE2CE6A} = {785A6394-7DDC-42FB-A7ED-2898EDC177EE}
 		{F403ED3B-EA47-415B-B9C9-4BA159B063D8} = {785A6394-7DDC-42FB-A7ED-2898EDC177EE}
 		{ACE8714C-1DDE-4D4D-B947-092ACB940888} = {5865CEC5-B728-447A-824B-4143CCC69C30}
 		{17253703-7209-43B3-863B-0A8FE5D524EB} = {5865CEC5-B728-447A-824B-4143CCC69C30}
@@ -177,5 +175,7 @@ Global
 		{08EBFFE9-3FBC-4778-B56F-EC6212BE743E} = {5865CEC5-B728-447A-824B-4143CCC69C30}
 		{DFDDC18D-7CDB-4842-94B5-5F7332AC95EC} = {5865CEC5-B728-447A-824B-4143CCC69C30}
 		{94B05EC3-2FE9-42D9-B765-A322DF2C87CD} = {5865CEC5-B728-447A-824B-4143CCC69C30}
+		{1A6B1B96-89C0-46A7-A6BF-3D43298D7E26} = {785A6394-7DDC-42FB-A7ED-2898EDC177EE}
+		{386FEAE8-9910-4816-B846-06633983F3B2} = {785A6394-7DDC-42FB-A7ED-2898EDC177EE}
 	EndGlobalSection
 EndGlobal
